+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - alias \\d to [0-9] in regex masks
 - clear masktemplate before submitting regardsless of the clearMaskOnLostFocus option
 
+### Fixed
+- min and max not working in datetime for 12 Hours format i.e. dd/mm/yyyy hh:MM:ss TT #2297
+
 ## [5.0.6 - 01/06/2021]
 
 ### Addition